api_spacex/requirements.txt
Django==2.2.6
django-filter==2.2.0
djangorestframework==3.10.3
requests==2.22.0
django-cors-headers==3.1.1
